requirements.txt
pymupdf
pytesseract
pdf2image
pillow
